Move factory to crawler/factory.
