+++ crawler/factory/factory.go
@@ -1,4 +1,4 @@
-package crawlworker
+package factory
 
 import (
 	"context"
